Skip a duplicate commit_push identity when HEAD is already recorded.
A clean-tree finalisation was appending a second checkpoint for the same SHA, so a settled resume failed the owned-span check before review. Extract the publisher collaborator so the logical type stays under the line ceiling.
