+++ runtime-kotlin/runtime-application/src/main/kotlin/skillbill/application/featuretask/FeatureTaskRuntimeSubtaskFinalisationCommitSupport.kt
@@ -1,10 +1,7 @@
 package skillbill.application.featuretask
 
-import skillbill.application.featuretask.model.FeatureTaskRuntimeCheckpointRefPruneRequest
 import skillbill.application.featuretask.model.FeatureTaskRuntimeSubtaskFinalisationBlocked
 import skillbill.application.featuretask.model.FeatureTaskRuntimeSubtaskFinalisationResult
-import skillbill.application.featuretask.model.FeatureTaskRuntimeSubtaskFinaliseRequest
-import skillbill.application.featuretask.model.FeatureTaskRuntimeSubtaskFinalised
 import skillbill.ports.workflow.gitops.model.WorkflowGitOperationResult
 import skillbill.ports.workflow.gitops.restoreIndexState
 
@@ -51,128 +48,7 @@ internal fun FeatureTaskRuntimeSubtaskFinalisation.restoreForeignFinalisationInd
 
 internal fun FeatureTaskRuntimeSubtaskFinalisation.commitAndPush(
   input: FinalisationCommitRequest,
-): FeatureTaskRuntimeSubtaskFinalisationResult {
-  val request = input.request
-  val stageable = input.stageable
-  val excluded = input.excluded
-  val restoreState = input.restoreState
-  val foreignStagedPaths = input.foreignStagedPaths
-  val foreignSnapshot = input.foreignSnapshot
-  val branch = request.metadata.branch
-  if (stageable.isEmpty()) {
-    val foreignRestored = restoreForeignIndex(foreignStagedPaths, foreignSnapshot)
-    if (foreignRestored != null) return blocked(foreignRestored)
-    return publishExistingHead(request, branch, excluded)
-  }
-  val decision = decide(
-    branch = branch,
-    identity = request.identity,
-    durableCommitSha = request.durableCommitSha,
-    sequenceNumber = request.sequenceNumber,
-  )
-  val rewrites = decision is FeatureTaskRuntimeSubtaskCommitAmend
-  val message = FeatureTaskRuntimeCheckpointMessage.finalise(
-    request.handoff.outcomeMessage,
-    request.metadata,
-    request.identity,
-  )
-  val commit = gitOperations.writeSubtaskCommitPreservingHistory(
-    SubtaskCommitPreservationRequest(
-      repoRoot = repoRoot,
-      decision = decision,
-      identity = request.identity,
-      message = message,
-      allowUnchangedIndex = true,
-      ownedPaths = stageable,
-      record = record,
-    ),
-  )
-  val commitSha = when (val outcome = finalisationCommitSha(commit, stageable, restoreState)) {
-    is FinalisationCommitShaBlocked -> return blocked(
-      restoreForeignFinalisationIndex(outcome.reason, foreignStagedPaths, foreignSnapshot),
-    )
-    is FinalisationCommitShaReady -> outcome.value
-  }
-  val foreignRestored = restoreForeignIndex(foreignStagedPaths, foreignSnapshot)
-  if (foreignRestored != null) return blocked(foreignRestored)
-  val recordFailure = recordCommit(commitSha, stageable)
-  return if (recordFailure != null) {
-    FeatureTaskRuntimeSubtaskFinalisationBlocked(recordFailure)
-  } else {
-    finalizeCommittedSubtask(
-      FinalizeCommittedSubtaskInput(
-        request = request,
-        branch = branch,
-        stageable = stageable,
-        excluded = excluded,
-        commitSha = commitSha,
-        rewrites = rewrites,
-      ),
-    )
-  }
-}
-
-private fun FeatureTaskRuntimeSubtaskFinalisation.publishExistingHead(
-  request: FeatureTaskRuntimeSubtaskFinaliseRequest,
-  branch: String,
-  excluded: List<String>,
-): FeatureTaskRuntimeSubtaskFinalisationResult {
-  val head = gitOperations.headCommitSha(repoRoot)
-  val commitSha = head.value.orEmpty().trim()
-  if (!head.ok || commitSha.isBlank()) {
-    return blocked("HEAD could not be resolved (${head.error})")
-  }
-  val recordFailure = recordCommit(commitSha, emptyList())
-  return if (recordFailure != null) {
-    FeatureTaskRuntimeSubtaskFinalisationBlocked(recordFailure)
-  } else {
-    finalizeCommittedSubtask(
-      FinalizeCommittedSubtaskInput(
-        request = request,
-        branch = branch,
-        stageable = emptyList(),
-        excluded = excluded,
-        commitSha = commitSha,
-        rewrites = false,
-      ),
-    )
-  }
-}
-
-private data class FinalizeCommittedSubtaskInput(
-  val request: FeatureTaskRuntimeSubtaskFinaliseRequest,
-  val branch: String,
-  val stageable: List<String>,
-  val excluded: List<String>,
-  val commitSha: String,
-  val rewrites: Boolean,
-)
-
-private fun FeatureTaskRuntimeSubtaskFinalisation.finalizeCommittedSubtask(
-  input: FinalizeCommittedSubtaskInput,
-): FeatureTaskRuntimeSubtaskFinalisationResult {
-  val forcedWithLease = input.rewrites && remoteDiverged(input.branch, input.commitSha)
-  val pushFailure = push(input.branch, input.request.identity, input.commitSha, forcedWithLease)
-  if (pushFailure != null) return blocked(pushFailure)
-  if (!input.request.manifestCommitSha.isNullOrBlank()) {
-    gitOperations.pruneSubtaskCheckpointRefs(
-      repoRoot = repoRoot,
-      request = FeatureTaskRuntimeCheckpointRefPruneRequest(
-        issueKey = input.request.identity.issueKey,
-        subtaskId = input.request.identity.subtaskId,
-        manifestCommitSha = input.request.manifestCommitSha,
-        featureBranch = input.branch,
-      ),
-      record = record,
-    )
-  }
-  return FeatureTaskRuntimeSubtaskFinalised(
-    commitSha = input.commitSha,
-    stagedPaths = input.stageable,
-    excludedSpecPaths = input.excluded,
-    forcedWithLease = forcedWithLease,
-  )
-}
+): FeatureTaskRuntimeSubtaskFinalisationResult = FeatureTaskRuntimeSubtaskCommitPublisher(this).commitAndPush(input)
 
 fun FeatureTaskRuntimeSubtaskFinalisation.restoring(error: String, paths: List<String>, snapshot: String): String {
   val restored = gitOperations.restoreIndexState(repoRoot, paths, snapshot)
